refactor: Car View<->Model 검증 로직을 나누기 위한 구조 변경
