+++ src/test/java/racingcar/domain/CarTest.java
@@ -9,12 +9,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class CarTest {
-
     @ParameterizedTest
     @ValueSource(ints = {4, 5})
     @DisplayName("자동차 이동시 랜덤값이 4이상인 경우 position변화 테스트")
     public void drive_forward_value(int value) {
-        Car car = new Car("is2js");
+        Car car = Car.fromName("is2js");
         car.drive(value);
 
         assertThat(1).isEqualTo(car.getPosition());
@@ -24,7 +23,7 @@ class CarTest {
     @ValueSource(ints = {1, 3})
     @DisplayName("자동차 이동시 랜덤값이 3이하인 경우 position변화 테스트")
     public void drive_stop_value(int value) {
-        Car car = new Car("is2js");
+        Car car = Car.fromName("is2js");
         car.drive(value);
 
         assertThat(0).isEqualTo(car.getPosition());
@@ -34,13 +33,13 @@ class CarTest {
     @DisplayName("자동차 이름이 유효한 길이인지 검사")
     @ValueSource(strings = {"5자이하", "is2js", "코니"})
     void valid_carName_length(String name) {
-        assertDoesNotThrow(() -> new Car(name));
+        assertDoesNotThrow(() -> Car.fromName(name));
     }
 
     @ParameterizedTest
     @DisplayName("자동차 이름이 유효하지 않은 길이인지 검사")
     @ValueSource(strings = {"5자초과닉넴", "is2js3js", "코니포코포비"})
     void invalid_carName_length(String name) {
-        assertThatThrownBy(() -> new Car(name)).isInstanceOf(IllegalArgumentException.class);
+        assertThatThrownBy(() -> Car.fromName(name)).isInstanceOf(IllegalArgumentException.class);
     }
 }
